+++ fasthtml/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.12.17"
+__version__ = "0.12.20"
 from .core import *