+++ fasthtml/fastapp.py
@@ -36,6 +36,7 @@ def fast_app(
         middleware:Optional[tuple]=None, # Standard Starlette middleware
         live:bool=False, # Enable live reloading
         debug:bool=False, # Passed to Starlette, indicating if debug tracebacks should be returned on errors
+        title:str="FastHTML page", # Default page title
         routes:Optional[tuple]=None, # Passed to Starlette
         exception_handlers:Optional[dict]=None, # Passed to Starlette
         on_startup:Optional[callable]=None, # Passed to Starlette
@@ -46,6 +47,7 @@ def fast_app(
         surreal:Optional[bool]=True, # Include surreal.js/scope headers?
         htmx:Optional[bool]=True, # Include HTMX header?
         exts:Optional[list|str]=None, # HTMX extension names to include
+        canonical:bool=True, # Automatically include canonical link?
         secret_key:Optional[str]=None, # Signing key for sessions
         key_fname:str='.sesskey', # Session cookie signing key file name
         session_cookie:str='session_', # Session cookie name
@@ -66,8 +68,8 @@ def fast_app(
     h = (picolink,) if pico or (pico is None and default_hdrs) else ()
     if hdrs: h += tuple(hdrs)
 
-    app = _app_factory(hdrs=h, ftrs=ftrs, before=before, middleware=middleware, live=live, debug=debug, routes=routes, exception_handlers=exception_handlers,
-                  on_startup=on_startup, on_shutdown=on_shutdown, lifespan=lifespan, default_hdrs=default_hdrs, secret_key=secret_key,
+    app = _app_factory(hdrs=h, ftrs=ftrs, before=before, middleware=middleware, live=live, debug=debug, title=title, routes=routes, exception_handlers=exception_handlers,
+                  on_startup=on_startup, on_shutdown=on_shutdown, lifespan=lifespan, default_hdrs=default_hdrs, secret_key=secret_key, canonical=canonical,
                   session_cookie=session_cookie, max_age=max_age, sess_path=sess_path, same_site=same_site, sess_https_only=sess_https_only,
                   sess_domain=sess_domain, key_fname=key_fname, exts=exts, surreal=surreal, htmx=htmx, htmlkw=htmlkw,
                   reload_attempts=reload_attempts, reload_interval=reload_interval, body_wrap=body_wrap, nb_hdrs=nb_hdrs, **(bodykw or {}))